requirements/prod.txt
asgiref==3.8.1
Django==4.2.16
django-cleanup==9.0.0
pillow==11.0.0
python-dotenv==1.0.1
sorl-thumbnail==12.11.0
sqlparse==0.5.1
tzdata==2024.2